s使用无锁队列

+++ JobSystem/WorkThreadStealQueue.h
@@ -9,8 +9,8 @@ class WorkThreadStealQueue {
 private:
 	std::mutex mutex;
 	Job* m_jobs[MAX_NUMBER_OF_JOBS_PERTTHREAD];
-	int m_bottom;
-	int m_top;
+	long m_bottom;
+	long m_top;
 public:
 	void Push(Job* job);
 	Job* Pop();
